--- Assets/Prefabs/Assembly/FourDoorsRoomC.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 07422990a956da54db9c32fd580f794e
-timeCreated: 1505148291
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
