--- Assets/Prefabs/Assembly/OneDoorRoom.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0783dce05b19919479cacdb1f8658c03
-timeCreated: 1502796685
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
