--- Assets/Prefabs/Assembly/RoofFoorDoors.prefab.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 59d225c8fa5954040b9e0801cee46c97
-timeCreated: 1508855428
-licenseType: Free
-NativeFormatImporter:
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
