--- Assets/Prefabs/Assembly/RoofOneDoor.prefab.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: e6043aaa39c771943aeffbee3d75dc88
-timeCreated: 1508855451
-licenseType: Free
-NativeFormatImporter:
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
