--- Assets/Prefabs/Assembly/ThreeDoorsRoom.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0861c5360ffaeaf4895560342e4f230b
-timeCreated: 1503567844
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
