--- Assets/Prefabs/Assembly/TwoDoorsRoomC.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 66c3af08eaf61b349b519e06800a50cf
-timeCreated: 1503567736
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
