--- Assets/Prefabs/Assembly/TwoDoorsRoomT.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: eac236f856801274a8458c35a07e2521
-timeCreated: 1502796904
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
