+++ Assets/Scripts/Dungeon/Dungeon/EndRoomBehaviour.cs
@@ -1,0 +1,7 @@
+﻿/** EndRoomBehaviour, public class
+ * This class is attached to the EndRoom of the Dungeon
+ **/
+public class EndRoomBehaviour : RoomBase
+{
+
+}
