+++ Assets/Scripts/Dungeon/Dungeon/StartRoomBehaviour.cs
@@ -1,0 +1,7 @@
+﻿/** StartRoomBehaviour, public class
+ * This class is attached to the StartRoom of the Dungeon
+ **/
+public class StartRoomBehaviour : RoomBase
+{
+
+}
