--- Assets/Scripts/Dungeon/RoomBase.cs
@@ -1,11 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public abstract class RoomBase : MonoBehaviour {
-
-    public Transform RoomNorth;
-    public Transform RoomSouth;
-    public Transform RoomEast;
-    public Transform RoomWest;
-}
