--- Assets/Scripts/Editor/MapEditor.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 5942e9a095d5a214bb911006c9e30f8c
-timeCreated: 1503564662
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
