--- Assets/Scripts/Generator/BoardCreator.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: c359798f49d8ff64aa94f387a4e0c77c
-timeCreated: 1504031309
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
