--- Assets/Scripts/Generator/Corridor.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 77a0a95280fee974790b86561a7b5396
-timeCreated: 1504031347
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
