--- Assets/Scripts/Generator/Room.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 5bea6aaeddc77c24499987c6dbf0e675
-timeCreated: 1504031335
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
