--- Assets/Scripts/Generator/oldGeneratorMap.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 6c8712ee289d84e41acf1b63e8ac1c1b
-timeCreated: 1517496503
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
